+++ R/data.R
@@ -21,11 +21,29 @@ NULL
 
 
 #' @docType data
-#' @title Sample dataset
-#' @name contrast_example
+#' @title More puppy therapy data
+#' @name puppy_love
 #' @keywords data
 #'
-#' @description Just for internal testing
+#' @description Fictitious data related to whether puppy therapy works when you
+#' adjust for a person’s love of puppies, taken from the `{discovr}` package
+#' (Field 2025)
+#'
+#' @details Following variables are included in the dataset:
+#' - `id``: Participant id
+#' - `dose`: Treatment group to which the participant was randomly assigned (No
+#'   puppies (control), 15 minutes of puppy therapy, 30 minutes of puppy
+#'   therapy)
+#' - `happiness`: Self-reported happiness from 0 (as unhappy as I can possibly
+#'   imagine being) to 10 (as happy as I can possibly imagine being)
+#' - `puppy_love`: Self-reported love of puppies from 0 (I am a weird person who
+#'   hates puppies, please be deeply suspicious of me) to 7 (puppies are the
+#'   best thing ever, one day I might marry one)
+#'
+#' For further details, see `?discovr::puppy_love`.
+#'
+#' @references Field, A. P. (2025). Discovering statistics using R and RStudio
+#' (2nd ed.). London: Sage.
 NULL