+++ R/estimate_predicted.R
@@ -130,6 +130,18 @@
 #' names matching the model frame (see [insight::get_data()]). This can be used
 #' to generate model predictions for specific combinations of predictor values.
 #'
+#' @section Finite mixture models:
+#'
+#' For finite mixture models (currently, only the [`brms::mixture()`] family
+#' from package *brms* is supported), use `predict = "classification"` with
+#' `data = NULL` to predict the class membership for each observation (e.g.,
+#' `estimate_prediction(model, predict = "classification")`). To return
+#' predicted values stratified by class membership, use `predict = "link"`
+#' (possibly in combination with `data` or `by`, e.g.
+#' `estimate_link(model, by = "predictor")`). Other `predict` options will
+#' return predicted values of the outcome for the full data, not stratified by
+#' class membership.
+#'
 #' @note
 #'
 #' These functions are built on top of [insight::get_predicted()] and correspond
@@ -193,7 +205,7 @@
 #' class `"estimate_predicted"`. Methods for [`visualisation_recipe()`][visualisation_recipe.estimate_predicted]
 #' and [`plot()`][visualisation_recipe.estimate_predicted] are available.
 #'
-#' @examplesIf all(insight::check_if_installed(c("see", "lme4", "rstanarm"), quietly = TRUE))
+#' @examplesIf all(insight::check_if_installed(c("see", "glmmTMB", "rstanarm"), quietly = TRUE))
 #' library(modelbased)
 #'
 #' # Linear Models
@@ -221,9 +233,17 @@
 #' estimate_relation(model)
 #'
 #' # Mixed models
-#' model <- lme4::lmer(mpg ~ wt + (1 | gear), data = mtcars)
+#' data(mtcars)
+#' mtcars$gear <- as.factor(mtcars$gear)
+#' model <- glmmTMB::glmmTMB(mpg ~ wt + (1 | gear), data = mtcars)
 #' estimate_expectation(model)
 #' estimate_relation(model)
+#'
+#' # Predict random effects and calculate contrasts
+#' estim <- estimate_relation(model, by = "gear")
+#' estim
+#'
+#' estimate_contrasts(estim)
 #'
 #' # Bayesian models
 #' \donttest{
@@ -353,6 +373,11 @@ estimate_relation <- function(model,
                                 iterations = NULL,
                                 keep_iterations = FALSE,
                                 ...) {
+  # return early for htest
+  if (inherits(model, "htest")) {
+    return(insight::get_predicted(model, ...))
+  }
+
   # only "by" or "data", but not both
   if (!is.null(by) && !is.null(data)) {
     insight::format_error("You can only specify one of `by` or `data`, but not both.")
@@ -476,6 +501,9 @@ estimate_relation <- function(model,
   predictions <- do.call(insight::get_predicted, c(prediction_args, dots))
   out <- as.data.frame(predictions, keep_iterations = keep_iterations)
 
+  # sanity check - did method return standard errors?
+  .check_standard_errors(out = out, model = model, ...)
+
   # select columns to copy - we don't want duplicates from the data grid
   columns_to_copy <- setdiff(colnames(data), colnames(out))
   if (length(columns_to_copy)) {
@@ -523,6 +551,7 @@ estimate_relation <- function(model,
   attr(out, "table_title") <- c("Model-based Predictions", "blue")
   attr(out, "coef_name") <- "Predicted"
   attr(out, "model_info") <- insight::model_info(model, response = 1)
+  attr(out, "predict") <- predict
   attr(out, "table_footer") <- .table_footer(
     out,
     by = grid_specs$at,