+++ tests/testthat/test-attributes_estimatefun.R
@@ -20,7 +20,7 @@ test_that("attributes_means", {
     c(
       "names", "class", "row.names", "at", "by", "focal_terms", "adjusted_for",
       "predict", "estimate", "transform", "datagrid", "preserve_range",
-      "model_info", "keep_iterations", "joint_test", "table_title",
+      "model_info", "keep_iterations", "joint_test", "vcov", "table_title",
       "table_footer", "model", "response", "ci", "backend", "coef_name"
     )
   )
@@ -48,7 +48,7 @@ test_that("attributes_means, contrasts", {
       "model", "response", "ci", "p_adjust", "backend", "focal_terms",
       "adjusted_for", "predict", "comparison", "contrast", "estimate",
       "transform", "datagrid", "preserve_range", "coef_name", "model_info",
-      "keep_iterations", "joint_test"
+      "keep_iterations", "joint_test", "vcov"
     )
   )
   estim <- suppressMessages(estimate_contrasts(model, "Species=c('setosa','virginica')", backend = "marginaleffects"))
@@ -59,7 +59,7 @@ test_that("attributes_means, contrasts", {
       "model", "response", "ci", "p_adjust", "backend", "focal_terms",
       "adjusted_for", "predict", "comparison", "contrast", "estimate",
       "transform", "datagrid", "preserve_range", "coef_name", "model_info",
-      "keep_iterations", "joint_test"
+      "keep_iterations", "joint_test", "vcov"
     )
   )
   estim <- suppressMessages(estimate_contrasts(model, "Species=c('setosa','virginica')", backend = "marginaleffects", estimate = "average"))
@@ -70,7 +70,7 @@ test_that("attributes_means, contrasts", {
       "model", "response", "ci", "p_adjust", "backend", "focal_terms",
       "adjusted_for", "predict", "comparison", "contrast", "estimate",
       "transform", "datagrid", "preserve_range", "coef_name", "model_info",
-      "contrast_filter", "keep_iterations", "joint_test"
+      "contrast_filter", "keep_iterations", "joint_test","vcov"
     )
   )
 })
@@ -95,7 +95,7 @@ test_that("attributes_means, slopes", {
     c(
       "names", "class", "row.names", "trend", "comparison", "p_adjust",
       "transform", "coef_name", "slope", "ci", "model_info", "keep_iterations",
-      "table_title", "table_footer", "model", "response"
+      "vcov", "table_title", "table_footer", "model", "response"
     )
   )
 })
@@ -114,8 +114,8 @@ test_that("attributes_means", {
     c(
       "names", "class", "row.names", "ci", "keep_iterations", "response",
       "transform", "model", "datagrid", "focal_terms", "preserve_range",
-      "table_title", "coef_name", "model_info", "table_footer", "adjusted_for",
-      "at_specs", "at", "by", "reference", "data"
+      "table_title", "coef_name", "model_info", "predict", "table_footer",
+      "adjusted_for", "at_specs", "at", "by", "reference", "data"
     )
   )
 })